refactor: split out dicom quick info
